+++ application/tank.h
@@ -3,7 +3,7 @@
 #define FUNCTIONS_H_INCLUDED
 #include"structs.h"
 #endif
-double isTankFull (hydrogenTank* tank);
+int isTankFull (hydrogenTank* tank);
 double tankFreeSpace(hydrogenTank* tank);
 int isValidIncreaseOfHydrogen(hydrogenTank* tank, double amountIncrease);
 double convertElectricityToHydrogen(double electricityMWh);
